pyproject.toml
[project]
name = "accomplish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
